refact: move shared functions into a shared module
Instead of importing from `upgrade_python_package`, move over critical
functions that are used in multiple modules into `utils.py`. Now we
should be independent from importing from `upgrade_python_package`.
